Bug/required and numbers

+++ app/Models/FormBuilding/NumberInputFormElement.php
@@ -34,7 +34,7 @@ class NumberInputFormElement extends Model
     protected $attributes = [
         'hideLabel' => false,
         'step' => 1,
-        'formatStyle' => 'decimal',
+        'formatStyle' => 'integer',
     ];
 
     /**
